pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gitstow"
version = "0.1.0"
description = "A git repository library manager — clone, organize, and maintain collections of repos you learn from"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Rish Madaan"},
]
keywords = ["git", "repository", "manager", "cli", "open-source", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
]

dependencies = [
    "typer[all]>=0.9",
    "pyyaml>=6.0",
    "rich>=13.0",
    "beaupy>=3.0",
]

[project.urls]
Repository = "https://github.com/rishmadaan/gitstow"
Issues = "https://github.com/rishmadaan/gitstow/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
gitstow = "gitstow.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/gitstow"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
